flesh out driver support and motor API

+++ dubcar_ws/src/vesc_driver/include/vesc_driver/vesc_packet.h
@@ -18,7 +18,7 @@ class VescPacket
     static const byte_t LARGE_PAYLOAD;
     static const byte_t SMALL_PAYLOAD;
     static const byte_t STOP_BYTE;
-    static const uint16_t LOWER_BYTE_MASK;
+    static const uint8_t LOWER_BYTE_MASK;
 
     VescPacket(const std::string &name, const Buffer &payload);
     const Buffer getBuffer() const;
